requirements.txt
django==2.0.5
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-registration-redux==2.4
django-dual-authentication==1.0
django-cron==0.5.0
django-dbbackup==3.2.0
django-debug-toolbar
